moved routing state to the state file
