+++ projects/core/src/routing/facade/routing.service.spec.ts
@@ -5,14 +5,13 @@ import { of } from 'rxjs';
 import { PageType } from '../../model/cms.model';
 import { SemanticPathService } from '../configurable-routes/url-translation/semantic-path.service';
 import { PageContext } from '../models/page-context.model';
-import * as fromStore from '../store';
-import { RouterState } from '../store/reducers/router.reducer';
+import * as fromStore from '../store/index';
 import { RoutingSelector } from '../store/selectors/index';
 import { RoutingService } from './routing.service';
 import createSpy = jasmine.createSpy;
 
 describe('RoutingService', () => {
-  let store: Store<RouterState>;
+  let store: Store<fromStore.RouterState>;
   let service: RoutingService;
   let urlService: SemanticPathService;
