+++ projects/core/src/routing/facade/routing.service.ts
@@ -7,7 +7,7 @@ import { SemanticPathService } from '../configurable-routes/url-translation/sema
 import { UrlCommands } from '../configurable-routes/url-translation/url-command';
 import { PageContext } from '../models/page-context.model';
 import * as fromStore from '../store';
-import { RouterState } from '../store/reducers/router.reducer';
+import { RouterState } from '../store/roting-state';
 import { RoutingSelector } from '../store/selectors/index';
 
 @Injectable({
