+++ projects/core/src/routing/store/reducers/router.reducer.ts
@@ -1,15 +1,15 @@
 import { InjectionToken, Provider } from '@angular/core';
-import { Params, RouterStateSnapshot } from '@angular/router';
+import { RouterStateSnapshot } from '@angular/router';
 import * as fromNgrxRouter from '@ngrx/router-store';
 import { ActionReducerMap } from '@ngrx/store';
 import { PageType } from '../../../model/cms.model';
 import { CmsActivatedRouteSnapshot } from '../../models/cms-route';
 import { PageContext } from '../../models/page-context.model';
-
-export interface RouterState
-  extends fromNgrxRouter.RouterReducerState<ActivatedRouterStateSnapshot> {
-  nextState?: ActivatedRouterStateSnapshot;
-}
+import {
+  ActivatedRouterStateSnapshot,
+  RouterState,
+  State,
+} from '../roting-state';
 
 export const initialState: RouterState = {
   navigationId: 0,
@@ -24,18 +24,6 @@ export const initialState: RouterState = {
   },
   nextState: undefined,
 };
-
-export interface ActivatedRouterStateSnapshot {
-  url: string;
-  queryParams: Params;
-  params: Params;
-  context: PageContext;
-  cmsRequired: boolean;
-}
-
-export interface State {
-  router: RouterState;
-}
 
 export function getReducers(): ActionReducerMap<State> {
   return {
