--- projects/core/src/routing/store/state.ts
@@ -1,1 +1,0 @@
-export const ROUTING_FEATURE = 'router';
